requirements.txt
# --- core runtime ---
fastapi==0.115.12
starlette==0.46.2
pydantic>=2.11.4
pydantic-settings>=2.3.2
python-magic>=0.4.27,<0.5.0
async_lru==1.0.1

# templating / docs
python-docx==1.1.2
docxtpl==0.19.1
lxml==5.4.0
PyYAML==6.0.2

# LLM + retries
openai==1.76.2
tenacity==9.1.2
httpx>=0.24.0,<0.28.0
anyio==4.9.0
httpcore==1.0.9
h11==0.16.0
sniffio==1.3.1

# file upload helpers
python-multipart==0.0.20
aiofiles==24.1.0

# PDF / image extraction
pdfplumber==0.11.6        # pulls pdfminer.six, pypdf
pillow==11.2.1
pytesseract==0.3.13
